Tests/LibWeb/Text/expected/cookie-cache-invalidation.txt

aggregate=aggregate=aggregate=base$one$two
